Restructure the entire resource management system
Resources exist on three different levels:
* GraphicsCache and SoundSystem contain actor agnostic references to animation
  and sound files. They also handle loading them off the hard disk.
* ResourceManager keeps a track record of the sets of resources any certain type
  of actor can have. Since actors' resource sets are loaded using the manifest,
  this caching level ensures there is no need to read the files over and over
  again as actors are recreated on the fly. Actor type specific overrides may
  also be added here - the lowest level keeps track of the defaults of each
  asset but some animations might not use the same settings.
* Each actor has a copy of the resource set they obtain from ResourceManager
  for easier local handling.

Due to high reliability on current animation and such, many aspects of the
engine ended up in a broken state at first. These should all be fixed now,
but it is possible the engine works differently in some areas than it did
earlier.

Ref #12: The animation regressions were fixed along with this update.
Ref #3: Caching is still done upon instantiation of the first instance
of a type of asset. Static method for this will probably require use of CRTP.