Implement a dedicated control handler
This ensures actors handle all key events at the same point of a
tick cycle instead of whenever the QKeyEvents fire, adding a layer
of consistency. It could also be used in the future to abstract
away the differences of keyboard and joystick inputs, though for
the moment, only keyboard will still be supported.

The key state registry only knows about the keys that are currently
held down, but it does keep track of all the keys that Qt's events
provide, so this should mean the support of more exotic keys should
not be a problem in the future.

Ref #11