Implement collapsible scenery
FPS parameter is not currently used; quick tests in JJ2 didn't
seem to provide any insight as to how it should affect the tiles,
other than making the animation through the frames instant if
it was set to anything other than 0.

As it is the last major navigation event to be implemented,
this fixes #10.