Fix next level when game started from an episode
Levels are now searched from the same episode as the previously
selected level. Moving to a level in a different episode is also
supported from a level configuration, though not currently
recommended - doing so will be confusing for regular levels and
there will be a dedicated way to do so between episodes (with
proper transition screens and such).