Split gameplay logic from the CarrotQt5 class
This huge commit removes gameplay related responsibilities
from the CarrotQt5 class, which is now more closely tied to
maintaining the general state of the program.

A new class, LevelManager, takes most of the gameplay global roles
that were previously done by CarrotQt5. All gameplay actors interact
with it and CarrotQt5 via another new class, ActorAPI. This interface
hides some implementation specifics of the level manager and provides
a cleaner and easier way for the actors to work with. Some low level
features still remain in the API, but they will be ironed out later on.

Ref #18